+++ tests/tests.py
@@ -16,12 +16,12 @@ class MyBackendTestCase(TestCase):
         else:
             return self.backend.authenticate(None, username, password)
 
-    def test_empty_role(self):
+    def test_role_without_groups_and_permissions(self):
         user = self.authenticate('user1', 'test')
         permissions = self.backend.get_all_permissions(user)
         self.assertSetEqual(permissions, set())
 
-    def test_role(self):
+    def test_role_with_groups_and_without_permissions(self):
         user = self.authenticate('user2', 'test')
         permissions = self.backend.get_all_permissions(user)
         self.assertSetEqual(permissions, {
@@ -31,6 +31,26 @@ class MyBackendTestCase(TestCase):
             'app.can_add_model2',
             'app.can_update_model2',
             'app.can_delete_model2',
+        })
+
+    def test_role_with_groups_and_permissions(self):
+        user = self.authenticate('user4', 'test')
+        permissions = self.backend.get_all_permissions(user)
+        self.assertSetEqual(permissions, {
+            'app.can_add_model1',
+            'app.can_update_model1',
+            'app.can_delete_model1',
+            'app.can_add_model2',
+            'app.can_update_model2',
+            'app.can_delete_model2',
+            'app.some_operation_model2',
+        })
+
+    def test_role_without_groups_and_with_permissions(self):
+        user = self.authenticate('user5', 'test')
+        permissions = self.backend.get_all_permissions(user)
+        self.assertSetEqual(permissions, {
+            'app.some_operation_model2',
         })
 
     def test_unauthorized(self):